clean up hooks
