+++ ts/src/components/Sidebar/ActionHistory/index.tsx
@@ -1,16 +1,13 @@
 import React from "react";
 import { useSelector } from "react-redux";
 
+import { usePlayerNames } from "../../../hooks/usePlayerNames";
 import { State } from "../../../state/domain";
 import History from "./History";
 
 const ActionHistory = () => {
-  const { history, playerNames } = useSelector(
-    ({ history, playerNames }: State) => ({
-      history,
-      playerNames,
-    })
-  );
+  const [playerNames] = usePlayerNames();
+  const history = useSelector(({ history }: State) => history);
 
   const mapped = history.map(({ action, timestamp, details }, i) => ({
     timestamp,
