+++ ts/src/components/Sidebar/Chat/ChatHistory.tsx
@@ -1,14 +1,13 @@
 import React from "react";
 import { useSelector } from "react-redux";
 
+import { usePlayerNames } from "../../../hooks/usePlayerNames";
 import { State } from "../../../state/domain";
 import History from "../ActionHistory/History";
 
 const ChatHistory = () => {
-  const { chat, playerNames } = useSelector(({ chat, playerNames }: State) => ({
-    chat,
-    playerNames,
-  }));
+  const [playerNames] = usePlayerNames();
+  const chat = useSelector(({ chat }: State) => chat);
 
   const mapped = chat.map(({ timestamp, clientId, message }) => ({
     timestamp,
