+++ ts/src/hooks/usePlayerNames.ts
@@ -1,0 +1,8 @@
+import { useSelector } from "react-redux";
+
+import { State } from "../state/domain";
+
+export const usePlayerNames = (): [Record<number, string>] => {
+  const playerNames = useSelector(({ playerNames }: State) => playerNames);
+  return [playerNames];
+};
